[Core][DbMetaData] - Moving fetchColumns method from TableMetadata class to Database class
